fix bad type

+++ my_parser.py
@@ -153,7 +153,7 @@ thrown, list of avaliable layers will be displayed.') ##**
     parser.add_argument("--replace_layer", action='store_true', help='Replace layer. For now replace ReLu to GaussA.') ##**
     parser.add_argument("--ll_scaling", type=float, default=0.01, help='Scaling for layer loss.') ##**
     parser.add_argument("--use_grad_pruning_at", type=int, help='Use gradient pruning at.') ##**
-    parser.add_argument("--grad_pruning_percent", type=int, default=0.01,
+    parser.add_argument("--grad_pruning_percent", type=float, default=0.01,
         help='Percent of gradient pruning neurons at given layer. Selected by std descending.') ##**
